requirements.txt
flask==3.0.3
google-generativeai==0.8.3
ollama==0.3.3
langchain-community==0.3.2
python-dotenv==1.0.1
openai==1.51.0  # optional fallback
anthropic>=0.30.0

